requirements.txt
# Document Processing
docling>=2.0.0

# Vector Database (Embedded File-Based)
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0
# Alternative: openai (if using OpenAI embeddings)
# openai>=1.0.0

# Google ADK Agent Library (Primary Querying)
# google-adk>=1.0.0  # Uncomment when available
# Note: Install from Google Cloud or official source when available

# Ollama (Local LLM for Specialized Tasks)
ollama>=0.1.0

# Web Search
duckduckgo-search>=4.0.0
# Alternative web search options:
# googlesearch-python>=1.2.0
# serpapi>=0.1.5

# Google Cloud (Healthcare Data Access)
google-cloud-bigquery>=3.0.0
google-cloud-spanner>=3.0.0
google-cloud-pubsub>=2.0.0
google-cloud-healthcare>=1.0.0

# FHIR Support
fhir.resources>=7.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.5.0

